requirements.txt
streamlit
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
pillow
beautifulsoup4